backend/requirements.txt
Flask==3.1.2
flask-cors==6.0.1
pandas==2.3.3
scikit-learn==1.7.2
joblib==1.5.2
easyocr==1.7.2
openai-whisper==20250625
numpy==2.2.6
scipy==1.15.3
torch
